[mod]重要！基础数据结构做了调整

+++ src/WtBtCore/MatchEngine.cpp
@@ -88,12 +88,12 @@ void MatchEngine::match_orders(WTSTickData* curTick, OrderIDs& to_erase)
 				//如果价格相等,需要先看排队位置,如果价格不等说明已经全部被大单吃掉了
 				if (!ordInfo._positive && decimal::eq(price, ordInfo._limit))
 				{
-					uint32_t& quepos = ordInfo._queue;
+					double& quepos = ordInfo._queue;
 
 					//如果成交量小于排队位置,则不能成交
 					if (volume <= quepos)
 					{
-						quepos -= (uint32_t)volume;
+						quepos -= volume;
 						continue;
 					}
 					else if (quepos != 0)
@@ -143,12 +143,12 @@ void MatchEngine::match_orders(WTSTickData* curTick, OrderIDs& to_erase)
 				//如果价格相等,需要先看排队位置,如果价格不等说明已经全部被大单吃掉了
 				if (!ordInfo._positive && decimal::eq(price, ordInfo._limit))
 				{
-					uint32_t& quepos = ordInfo._queue;
+					double& quepos = ordInfo._queue;
 
 					//如果成交量小于排队位置,则不能成交
 					if (volume <= quepos)
 					{
-						quepos -= (uint32_t)volume;
+						quepos -= volume;
 						continue;
 					}
 					else if (quepos != 0)
@@ -193,14 +193,14 @@ void MatchEngine::update_lob(WTSTickData* curTick)
 		uint32_t px = PRICE_DOUBLE_TO_INT(curTick->askprice(i));
 		if (px != 0)
 		{
-			uint32_t& volume = curBook._items[px];
+			double& volume = curBook._items[px];
 			volume = curTick->askqty(i);
 		}
 
 		px = PRICE_DOUBLE_TO_INT(curTick->bidprice(i));
 		if (px != 0)
 		{
-			uint32_t& volume = curBook._items[px];
+			double& volume = curBook._items[px];
 			volume = curTick->askqty(i);
 		}
 	}
